pyproject.toml
[project]
name = "gfdm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.1",
    "ema-pytorch>=0.7.7",
    "matplotlib>=3.10.1",
    "pytorch-lightning>=2.5.0.post0",
    "scipy>=1.15.2",
    "torchvision>=0.21.0",
    "wandb>=0.19.8",
]
